Add line attributes class
Instances of this class describe the visual attributes of lines. I believe grouping related properties using this class results in a cleaner API. Currently only colour and width are modelling, but more attributes could be added, and only the only interface than needs to change is the line attributes class. This is more scalable.